Round usage usage to 1 decimal place

+++ dto/v1/body/user.go
@@ -67,6 +67,6 @@ type Quota struct {
 type Usage struct {
 	CpuCores  float64 `json:"cpuCores"`
 	RAM       float64 `json:"ram"`
-	DiskSize  float64 `json:"diskSize"`
+	DiskSize  int     `json:"diskSize"`
 	Snapshots int     `json:"snapshots"`
 }
